backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.35
python-dotenv==1.0.0
httpx==0.28.0
pydantic==2.10.0
aiosqlite==0.20.0